Fix error in bindataRead
We might try to call gz.Close even if err is non-nil and we don't use
the result. Also return the correct error if we can't gz.Close().

Likely this hasn't been triggered in real life because these are all
in-memory data structures but still seems good to do the right thing.

Regenerate the output files with the new generated code.

Fixes jteeuwen/go-bindata#94.