Improve performance
Previously calling format.Source would take up most of the runtime.
But most of the file is formatted, we just want to use go/format
to whitespace align the tables per the Go rules. Calling go/format
on just this part of the file yields a significant performance
improvement by avoiding calling it on the body of the file.

Use a bytes.Buffer as a file argument instead of an io.Writer which
allows using WriteByte and WriteString in places. Avoid calling
fmt.Fprintf with no arguments.

    name       old time/op    new time/op    delta
    Bindata-4     558ms ± 1%     216ms ± 6%   -61.34%  (p=0.008 n=5+5)

    name       old speed      new speed      delta
    Bindata-4  12.3MB/s ± 1%  31.8MB/s ± 6%  +159.03%  (p=0.008 n=5+5)

    name       old alloc/op   new alloc/op   delta
    Bindata-4     353MB ± 0%     162MB ± 0%   -54.21%  (p=0.008 n=5+5)

    name       old allocs/op  new allocs/op  delta
    Bindata-4     46.1k ± 0%     15.7k ± 0%   -66.00%  (p=0.008 n=5+5)