convert: fix asset path bug when including files outside prefix
Before this patch, if you ran this command:

./go-bindata -o files.go -pkg files -prefix testdata/assets testdata/assets/* testdata/in/*

The asset paths would look like:

   bindata.go (from testdata/assets)
   testdata/in/a/test.asset
   testdata/in/b/test.asset
   testdata/in/c/test.asset
   testdata/in/test.asset/test.asset

For the test.asset file in testdata/in, it made a new directory with the
same name as the asset.

After this patch it looks like:

   bindata.go (from testdata/assets)
   testdata/in/a/test.asset
   testdata/in/b/test.asset
   testdata/in/c/test.asset
   testdata/in/test.asset

This is because when processing the files from the "in/*" glob it
treated every value as a directory and appended the file name, even
though test.asset is already a file. This only occurs when using a
prefix because when the prefix is empty we use relative paths and fall
back to using the original asset.Name.